requirements.txt
aiogram>=3.0
sqlalchemy[asyncio]
aiosqlite
pydantic-settings
python-dotenv
